gcp_hospital_voice_form-fix-multicallfix/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
channels==4.1.0
daphne==4.1.2
websockets==12.0
idna==3.10
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
